requirements.txt
appdirs==1.4.4
attrs==19.3.0
certifi==2020.4.5.1
click==7.1.2
cycler==0.10.0
decorator==4.4.2
flake8==3.8.3
future==0.18.2
imageio==2.8.0
joblib==0.15.1
kiwisolver==1.2.0
lmdb==0.98
matplotlib==3.2.1
mccabe==0.6.1
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
munch==2.5.0
networkx==2.4
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.34
pandas==1.0.4
pathspec==0.8.0
Pillow==7.1.2
pretrainedmodels==0.7.4
protobuf==3.12.2
pyarrow==0.17.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
regex==2020.6.8
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
tensorboardX==2.0
threadpoolctl==2.1.0
tifffile==2020.6.3
toml==0.10.1
torch==1.5.0
torchtoolbox==0.1.4.1
torchvision==0.6.0
tqdm==4.46.1
typed-ast==1.4.1
wincertstore==0.2
